\name{perspectev-internal}
\alias{perspectev.permute}
\alias{perspectev.simulation.permute}
\alias{mcpRange}
\alias{t1Range}
\alias{t2Range}
\alias{t1Var}
\alias{t2Var}
\alias{t1t2Covar}
\alias{t1Unique}
\alias{t2Unique}
\alias{t1t2Unique}
\alias{dnaTheta}

\title{Internal Functions}
\description{
 Internal perspectev functions
}
\details{
These are not meant to be called by the user except potentially as trait functions. See tutorial for details (http://evolve.zoo.ox.ac.uk/Evolve/Perspectev.html).
}
\keyword{ internal }
