\name{perspectev-package}
\alias{perspectev-package}
\alias{perspectev}
\docType{package}
\title{
Permutation of Species During Turnover Events
}
\description{
Provides a robust framework for analyzing the reducibility of differential survival with respect to higher level trait variation. In addition to its primary test, it also provides functions for simulation-based power analysis, reading in common data set formats, and visualizing results. For tutorial see: http://evolve.zoo.ox.ac.uk/Evolve/Perspectev.html.
}
\details{
\tabular{ll}{
Package: \tab perspectev\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-08-07\cr
License: \tab GPL-3\cr
}

}
\author{
Kenneth B. Hoehn <perspectev@gmail.com>
}

\keyword{ package }

